+++ packages/browser-wallet/src/shared/utils/token-helpers.ts
@@ -16,6 +16,7 @@ import {
     ContractName,
     EntrypointName,
     Energy,
+    CIS2Contract,
 } from '@concordium/web-sdk';
 import { MetadataUrl, TokenMetadata, TokenIdAndMetadata } from '@shared/storage/types';
 import { CIS2_SCHEMA_CONTRACT_NAME, CIS2_SCHEMA } from '@popup/constants/schema';
@@ -64,43 +65,6 @@ export function getMetadataParameter(ids: string[]): Buffer {
 }
 
 /**
- * Returns the url for the token metadata.
- * returnValue is assumed to be a HEX-encoded string.
- */
-function deserializeTokenMetadataReturnValue(returnValue: string): MetadataUrl[] {
-    const buf = Buffer.from(returnValue, 'hex');
-    const n = buf.readUInt16LE(0);
-    let cursor = 2; // First 2 bytes hold number of token amounts included in response.
-    const urls: MetadataUrl[] = [];
-
-    // eslint-disable-next-line no-plusplus
-    for (let i = 0; i < n; i++) {
-        const length = buf.readUInt16LE(cursor);
-        const urlStart = cursor + 2;
-        const urlEnd = urlStart + length;
-
-        const url = Buffer.from(buf.subarray(urlStart, urlEnd)).toString('utf8');
-
-        cursor = urlEnd;
-
-        const hasChecksum = buf.readUInt8(cursor);
-        cursor += 1;
-
-        if (hasChecksum === 1) {
-            const hash = Buffer.from(buf.subarray(cursor, cursor + 32)).toString('hex');
-            cursor += 32;
-            urls.push({ url, hash });
-        } else if (hasChecksum === 0) {
-            urls.push({ url });
-        } else {
-            throw new Error('Deserialization failed: boolean value had an unexpected value');
-        }
-    }
-
-    return urls;
-}
-
-/**
  * Confirms that the given smart contract instance is CIS-2 compliant
  */
 export async function confirmCIS2Contract(
@@ -124,36 +88,6 @@ export async function confirmCIS2Contract(
     }
 
     return undefined;
-}
-
-/**
- * Determines the metadata url for the given token.
- */
-export function getTokenUrl(
-    client: ConcordiumGRPCClient,
-    ids: string[],
-    { contractName, index, subindex }: ContractDetails
-): Promise<MetadataUrl[]> {
-    return new Promise((resolve, reject) => {
-        client
-            .invokeContract({
-                contract: ContractAddress.create(index, subindex),
-                method: ReceiveName.fromString(`${contractName}.tokenMetadata`),
-                parameter: Parameter.fromBuffer(getMetadataParameter(ids)),
-            })
-            .then((returnValue) => {
-                if (returnValue && returnValue.tag === 'success' && returnValue.returnValue) {
-                    try {
-                        resolve(deserializeTokenMetadataReturnValue(ReturnValue.toHexString(returnValue.returnValue)));
-                    } catch (e) {
-                        reject(e);
-                    }
-                } else {
-                    // TODO: perhaps we need to make this error more precise
-                    reject(new Error('Token does not exist in this contract'));
-                }
-            });
-    });
 }
 
 function confirmMetadataUrl(field?: MetadataUrl) {
@@ -254,26 +188,19 @@ export type ContractBalances = Record<string, bigint | undefined>;
 
 export const getContractBalances = async (
     client: ConcordiumGRPCClient,
-    index: bigint,
-    subindex: bigint,
+    contractDetails: ContractDetails,
     tokenIds: string[],
     accountAddress: string,
     onError?: (error: string) => void
 ): Promise<ContractBalances> => {
-    const instanceInfo = await client.getInstanceInfo(ContractAddress.create(index, subindex));
-
-    if (instanceInfo === undefined) {
-        return {};
-    }
-
     const result = await client.invokeContract({
-        contract: ContractAddress.create(index, subindex),
-        method: ReceiveName.fromString(`${instanceInfo.name.value.substring(5)}.balanceOf`),
+        contract: ContractAddress.create(contractDetails.index, contractDetails.subindex),
+        method: ReceiveName.fromString(`${contractDetails.contractName}.balanceOf`),
         parameter: Parameter.fromBuffer(serializeBalanceParameter(tokenIds, accountAddress)),
     });
 
     if (result === undefined || result.tag === 'failure' || result.returnValue === undefined) {
-        onError?.(`Failed to retrieve balances for index ${index.toString()}`);
+        onError?.(`Failed to retrieve balances for index ${contractDetails.index.toString()}`);
         return {};
     }
 
@@ -396,12 +323,14 @@ export async function getTokenTransferEnergy(
     return { execution, total };
 }
 
-type GetTokensResult = {
+type TokenData = {
     id: string;
     metadataLink: string;
     metadata: TokenMetadata | undefined;
     balance: bigint;
-}[];
+};
+
+type GetTokensResult = TokenData[];
 
 export async function getTokens(
     contractDetails: ContractDetails,
@@ -410,52 +339,43 @@ export async function getTokens(
     ids: string[],
     onError?: (error: string) => void
 ): Promise<GetTokensResult> {
-    const metadataPromise: Promise<[MetadataUrl[], Array<TokenMetadata | undefined>]> = (async () => {
-        let metadataUrls;
-        try {
-            metadataUrls = await getTokenUrl(client, ids, contractDetails);
-        } catch (e) {
-            onError?.(`Failed to get metadata urls on index: ${contractDetails.index}`);
-            return [[], []];
-        }
-        const metadata = await Promise.all(
-            metadataUrls.map((url, index) =>
-                getTokenMetadata(url).catch((error) => {
-                    onError?.(`id: "${ids[index]}": ${error.message}`);
-                    return Promise.resolve(undefined);
-                })
-            )
-        );
-
-        return [metadataUrls, metadata];
-    })();
-
-    const balancesPromise = getContractBalances(
+    const contract = new CIS2Contract(
         client,
-        contractDetails.index,
-        contractDetails.subindex,
-        ids,
-        account,
-        onError
+        ContractAddress.create(contractDetails.index, contractDetails.subindex),
+        ContractName.fromString(contractDetails.contractName)
+    );
+    const tokenData: (TokenData | undefined)[] = await Promise.all(
+        ids.map(async (id, index) => {
+            let metadataUrl;
+            try {
+                metadataUrl = await contract.tokenMetadata(id);
+            } catch (e) {
+                onError?.(`id: "${id}: Failed to get metadata url`);
+                return undefined;
+            }
+
+            let metadata;
+            try {
+                metadata = await getTokenMetadata(metadataUrl);
+            } catch (error) {
+                onError?.(`id: "${ids[index]}": ${(error as Error).message}`);
+                return undefined;
+            }
+
+            let balance: bigint;
+            try {
+                balance =
+                    (await contract.balanceOf({ address: AccountAddress.fromBase58(account), tokenId: id })) || 0n;
+            } catch (error) {
+                onError?.(`id: "${ids[index]}": Failed to get token balance`);
+                return undefined;
+            }
+
+            return { id, metadataLink: metadataUrl.url, metadata, balance };
+        })
     );
 
-    const [[metadataUrls, metadata], balances] = await Promise.all([metadataPromise, balancesPromise]); // Run in parallel.
-
-    return ids.reduce(
-        (acc, id, i) =>
-            id in balances && metadata[i]
-                ? [
-                      ...acc,
-                      {
-                          id,
-                          metadataLink: metadataUrls[i].url,
-                          metadata: metadata[i],
-                          balance: balances[id] ?? 0n,
-                      },
-                  ]
-                : acc,
-        [] as GetTokensResult
-    );
+    return tokenData.filter((data): data is TokenData => Boolean(data));
 }
 
 const MAX_SYMBOL_LENGTH = 10;