+++ packages/browser-wallet/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+-   CIS-2 token lookup no longer blocks choosing a contract, if looking up metadata or balance only fails for some tokens.
+
 ### Fixed
 
 -   Inject script loading wasm module, unnecessarily.